routes new et delete sécurisé

+++ src/Controller/ShopController.php
@@ -28,10 +28,10 @@ class ShopController extends AbstractController
         $shopUser = $user->getShop();
         $idShop = $shopUser->getId();
         $shop = $shopRepository->findOneBy(['id' => $idShop]);
-        dump ($idShop);
+        dump($idShop);
 
         dump($shop);
-        
+
         return $this->render('shop/index.html.twig', [
             'shop' => $shop,
         ]);
@@ -43,44 +43,51 @@ class ShopController extends AbstractController
     public function new(Request $request,  SluggerInterface $slugger): Response
     {
         $user = $this->getUser();
-        
+
         $shop = new Shop();
         $form = $this->createForm(ShopType::class, $shop);
         $form->handleRequest($request);
+        dump($user->getShop());
 
-        if ($form->isSubmitted() && $form->isValid()) {
-    
-            $img = $form->get('img')->getData();
-            if ($img) {
-                $originalFilename = pathinfo($img->getClientOriginalName(), PATHINFO_FILENAME);
-                // this is needed to safely include the file name as part of the URL
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $img->guessExtension();
+        if ($user->getShop() == null) {
 
-                // Move the file to the directory where brochures are stored
-                try {
-                    $img->move(
-                        $this->getParameter('photos_directory'),        // NE PAS OUBLIER DE CREER LE DOSSIER
-                        $newFilename
-                    );
-                } catch (FileException $e) {
-                   
+            if ($form->isSubmitted() && $form->isValid()) {
+
+                $img = $form->get('img')->getData();
+
+
+                if ($img) {
+                    $originalFilename = pathinfo($img->getClientOriginalName(), PATHINFO_FILENAME);
+                    // this is needed to safely include the file name as part of the URL
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $img->guessExtension();
+
+                    // Move the file to the directory where brochures are stored
+                    try {
+                        $img->move(
+                            $this->getParameter('photos_directory'),        // NE PAS OUBLIER DE CREER LE DOSSIER
+                            $newFilename
+                        );
+                    } catch (FileException $e) {
+                    }
+
+                    $shop->setImg($newFilename);       // ON ENREGISTRE LE NOM DU FICHIER
+                } else {
+                    $defaultImg = "Asset-1@2x-1.png";
+                    $shop->setImg($defaultImg);
                 }
+                $shop->setUser($user);
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($shop);
+                $entityManager->flush();
+                dump($shop);
 
-                $shop->setImg($newFilename);       // ON ENREGISTRE LE NOM DU FICHIER
-            } 
-            else {
-                $defaultImg = "Asset-1@2x-1.png";
-                $shop->setImg($defaultImg);
+                return $this->redirectToRoute('shop_index');
             }
-            $shop->setUser($user);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($shop);
-            $entityManager->flush();
-            dump($shop);
-
-           return $this->redirectToRoute('shop_index');
+        } else {
+            return $this->redirectToRoute('shop_index');
         }
+
 
         return $this->render('shop/new.html.twig', [
             'shop' => $shop,
@@ -123,10 +130,16 @@ class ShopController extends AbstractController
      */
     public function delete(Request $request, Shop $shop): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$shop->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($shop);
-            $entityManager->flush();
+        $user = $this->getUser();
+        if (in_array('ROLE_ADMIN', $user->getRoles())) {
+
+            if ($this->isCsrfTokenValid('delete' . $shop->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($shop);
+                $entityManager->flush();
+            }
+        } else {
+            return $this->redirectToRoute('shop_index');
         }
 
         return $this->redirectToRoute('shop_index');
